The highest mark in each subject and the roll no. of the student who obtained it.
Problem 04: The annual examination result of 20 students is tabulated as follow:
Roll no.              Sub1                 Sub2                      Sub3
Write a C program to read the data and determine the followings:
a)	Total marks obtained by each student
b)	The highest mark in each subject and the roll no. of the student who obtained it.
c)	The student who obtained the highest total marks.